refactor : vo 패키지 삭제
