+++ src/main/java/lotto/domain/Summary.java
@@ -1,4 +1,4 @@
-package lotto.vo;
+package lotto.domain;
 
 import java.util.Collections;
 import java.util.LinkedList;
