+++ src/main/java/lotto/domain/Ticket.java
@@ -1,4 +1,4 @@
-package lotto.vo;
+package lotto.domain;
 
 import java.util.Objects;
