+++ src/test/java/lotto/LotteryClerkTest.java
@@ -1,7 +1,7 @@
 package lotto;
 
-import lotto.vo.Ticket;
-import lotto.vo.Money;
+import lotto.domain.Ticket;
+import lotto.domain.Money;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
