+++ src/test/java/lotto/LotteryStoreTest.java
@@ -1,7 +1,7 @@
 package lotto;
 
-import lotto.vo.Lotteries;
-import lotto.vo.Money;
+import lotto.domain.Lotteries;
+import lotto.domain.Money;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
