+++ src/test/java/lotto/domain/MoneyTest.java
@@ -1,4 +1,4 @@
-package lotto.vo;
+package lotto.domain;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
