+++ src/test/java/lotto/domain/SummaryTest.java
@@ -1,4 +1,4 @@
-package lotto.vo;
+package lotto.domain;
 
 import org.junit.jupiter.api.Test;
